wait to run `mpdup`
On boot, the `music` script is ran which in turn runs the `mpdup` script. However, `mpd` (which is started through `.xprofile`) does not start immediately, and as a consequence `mpc` throws an error. Thus, we wait five seconds to give time for `mpd` to fully start. If a better alternative is found, it will be used.