fix moonphase and weather scripts
If logging in quickly, curl downloads an empty file as the network takes time to initialize, causing the script to output nothing to the statusbar. This commit checks for this case before properly retrieving information -- however, this does not solve the race condition between curl and the network and merely solves the case where the network has already initialized.